server/requirements.txt
Flask
Flask-Cors
python-dotenv
requests
tencentcloud-sdk-python 